Started control module
Final product will include a timer module
Frame control
Address control
SDRAM control module

Basically this is a start for using the SDRAM and controlling the
addresses.
I know How to impliment this to include individual pixel changing.
So i throughly thought this through.

Also i changed the SPI module so it will convert the 8 bit input and
store the data on a 16 bit buffer so it would make easy conversons for
the SDRAM.